whisper-api/requirements.txt
openai-whisper
torch
torchaudio
numpy
flask
googletrans==4.0.0-rc1
